Move board specific indiviudal components
