+++ apps/src/lib/kits/maker/boards/circuitPlayground/Switch.js
