+++ apps/src/lib/kits/maker/boards/microBit/Accelerometer.js
@@ -1,8 +1,5 @@
 import {EventEmitter} from 'events';
-import {
-  sensor_channels,
-  roundToHundredth
-} from './boards/microBit/MicroBitConstants';
+import {sensor_channels, roundToHundredth} from './MicroBitConstants';
 
 // Transfer the acceleration units from milli-g to meters/second^2
 // Round to the nearest hundredth
