+++ apps/src/lib/kits/maker/boards/microBit/MicroBitBoard.js
@@ -8,8 +8,8 @@ import {
   componentConstructors
 } from './MicroBitComponents';
 import MBFirmataWrapper from '../../MBFirmataWrapper';
-import ExternalLed from '../../ExternalLed';
-import ExternalButton from '../../ExternalButton';
+import ExternalLed from './ExternalLed';
+import ExternalButton from './ExternalButton';
 
 /**
  * Controller interface for BBC micro:bit board using
