+++ apps/test/unit/lib/kits/maker/AccelerometerTest.js
@@ -1,5 +1,5 @@
 import {expect} from '../../../../util/reconfiguredChai';
-import Accelerometer from '@cdo/apps/lib/kits/maker/Accelerometer';
+import Accelerometer from '@cdo/apps/lib/kits/maker/boards/microBit/Accelerometer';
 import {MicrobitStubBoard} from './makeStubBoard';
 import {sensor_channels} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitConstants';
 import sinon from 'sinon';
