+++ apps/test/unit/lib/kits/maker/LedMatrixTest.js
@@ -2,7 +2,7 @@
 import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import {MicrobitStubBoard} from './makeStubBoard';
-import LedMatrix from '@cdo/apps/lib/kits/maker/LedMatrix';
+import LedMatrix from '@cdo/apps/lib/kits/maker/boards/microBit/LedMatrix';
 
 describe('LedMatrix', function() {
   describe('on() and off()', () => {
