+++ apps/test/unit/lib/kits/maker/LedTest.js
@@ -3,7 +3,7 @@ import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
 import {makeStubBoard} from './makeStubBoard';
-import Led from '@cdo/apps/lib/kits/maker/Led';
+import Led from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/Led';
 
 describe('Led', function() {
   it('is a five.Led', () => {
