+++ apps/test/unit/lib/kits/maker/MicroBitBoardTest.js
@@ -5,8 +5,8 @@ import sinon from 'sinon';
 import {itImplementsTheMakerBoardInterface} from './MakerBoardTest';
 import _ from 'lodash';
 import {EXTERNAL_PINS} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitConstants';
-import ExternalLed from '@cdo/apps/lib/kits/maker/ExternalLed';
-import ExternalButton from '@cdo/apps/lib/kits/maker/ExternalButton';
+import ExternalLed from '@cdo/apps/lib/kits/maker/boards/microBit/ExternalLed';
+import ExternalButton from '@cdo/apps/lib/kits/maker/boards/microBit/ExternalButton';
 
 describe('MicroBitBoard', () => {
   let board;
