+++ apps/test/unit/lib/kits/maker/NeoPixelTest.js
@@ -2,7 +2,7 @@
 import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
-import NeoPixel from '@cdo/apps/lib/kits/maker/NeoPixel';
+import NeoPixel from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/NeoPixel';
 
 describe('NeoPixel', function() {
   beforeEach(() => {
