+++ apps/test/unit/lib/kits/maker/PiezoTest.js
@@ -2,7 +2,7 @@
 import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
-import Piezo from '@cdo/apps/lib/kits/maker/Piezo';
+import Piezo from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/Piezo';
 
 describe('Piezo', function() {
   beforeEach(function() {
