+++ apps/test/unit/lib/kits/maker/SwitchTest.js
@@ -4,7 +4,7 @@ import {EventEmitter} from 'events'; // provided by webpack's node-libs-browser
 import Switch, {
   READ_ONLY_PROPERTIES,
   READ_WRITE_PROPERTIES
-} from '@cdo/apps/lib/kits/maker/Switch';
+} from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/Switch';
 
 describe('Switch', () => {
   let fakeJohnnyFiveSwitch;
