+++ apps/test/unit/lib/kits/maker/TouchSensorTest.js
@@ -2,7 +2,7 @@
 import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import {EventEmitter} from 'events'; // provided by webpack's node-libs-browser
-import TouchSensor from '@cdo/apps/lib/kits/maker/TouchSensor';
+import TouchSensor from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/TouchSensor';
 
 const DOWN = 'down';
 const HOLD = 'hold';
